+++ packages/grid/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.32.0](https://github.com/deephaven/web-client-ui/compare/v0.31.1...v0.32.0) (2023-03-10)
+
+### Bug Fixes
+
+- Grid rendering header incorrectly when hiding all children in a group via layout hints ([#1139](https://github.com/deephaven/web-client-ui/issues/1139)) ([2fbccc6](https://github.com/deephaven/web-client-ui/commit/2fbccc60a7fe55264e7dceb260ba3962957a8eba)), closes [#1097](https://github.com/deephaven/web-client-ui/issues/1097)
+
+### Features
+
+- Add support for clickable links ([#1088](https://github.com/deephaven/web-client-ui/issues/1088)) ([f7f918e](https://github.com/deephaven/web-client-ui/commit/f7f918e7f0c5f1b0fb4030eb748010aaf4d196df)), closes [#712](https://github.com/deephaven/web-client-ui/issues/712)
+
+# [0.31.0](https://github.com/deephaven/web-client-ui/compare/v0.30.1...v0.31.0) (2023-03-03)
+
+**Note:** Version bump only for package @deephaven/grid
+
 # [0.30.0](https://github.com/deephaven/web-client-ui/compare/v0.29.1...v0.30.0) (2023-02-13)
 
 **Note:** Version bump only for package @deephaven/grid