+++ packages/jsapi-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.32.0](https://github.com/deephaven/web-client-ui/compare/v0.31.1...v0.32.0) (2023-03-10)
+
+**Note:** Version bump only for package @deephaven/jsapi-components
+
+# [0.31.0](https://github.com/deephaven/web-client-ui/compare/v0.30.1...v0.31.0) (2023-03-03)
+
+**Note:** Version bump only for package @deephaven/jsapi-components
+
 ## [0.30.1](https://github.com/deephaven/web-client-ui/compare/v0.30.0...v0.30.1) (2023-02-16)
 
 **Note:** Version bump only for package @deephaven/jsapi-components